third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-allowed.https-expected.txt

This is a testharness.js-based test.
FAIL Insecure resource loads inside Worker when 'AllowRunningInsecureContent' is set. assert_equals: expected "LOADED" but got "LOAD FAILED"
Harness: the test ran to completion.
